Refactor quizmodel sub-package to quiz
