+++ src/main/java/seedu/address/model/modelmanager/quiz/QuizModelManager.java
@@ -1,4 +1,4 @@
-package seedu.address.model.modelmanager.quizmodel;
+package seedu.address.model.modelmanager.quiz;
 
 import static seedu.address.commons.util.CollectionUtil.requireAllNonNull;
