+++ src/main/java/seedu/address/model/modelmanager/quiz/exceptions/NotInitialisedException.java
@@ -1,4 +1,4 @@
-package seedu.address.model.modelmanager.quizmodel.exceptions;
+package seedu.address.model.modelmanager.quiz.exceptions;
 
 /**
  *  Represents not yet initialised error encountered by Quiz
