+++ src/test/java/seedu/address/quiz/QuizTest.java
@@ -9,8 +9,8 @@ import java.util.List;
 
 import org.junit.Test;
 
-import seedu.address.model.modelmanager.quizmodel.Quiz;
-import seedu.address.model.modelmanager.quizmodel.QuizCard;
+import seedu.address.model.modelmanager.quiz.Quiz;
+import seedu.address.model.modelmanager.quiz.QuizCard;
 import seedu.address.testutil.Assert;
 
 public class QuizTest {
